Extract query classes for Participation and subclasses.
