+++ opengever/contact/models/participation.py
@@ -9,65 +9,18 @@ from opengever.contact.ogdsuser import OgdsUserToContactAdapter
 from opengever.ogds.base.utils import ogds_service
 from opengever.ogds.models import UNIT_ID_LENGTH
 from opengever.ogds.models import USER_ID_LENGTH
-from opengever.ogds.models.query import BaseQuery
 from sqlalchemy import Column
 from sqlalchemy import ForeignKey
 from sqlalchemy import Integer
-from sqlalchemy import or_
 from sqlalchemy import String
 from sqlalchemy.orm import composite
 from sqlalchemy.orm import relationship
 from sqlalchemy.schema import Sequence
 
 
-class ParticipationQuery(BaseQuery):
-
-    def by_dossier(self, dossier):
-        return self.filter_by(dossier_oguid=Oguid.for_object(dossier))
-
-    def by_organization(self, organization):
-        """Returns both ContactParticapations and OrgRoleParticipation
-        of the given organization.
-        """
-        return self._org_role_join().filter(
-            or_(OrgRole.organization == organization,
-                ContactParticipation.contact == organization))
-
-    def by_person(self, person):
-        """Returns both ContactParticapations and OrgRoleParticipation
-        of the given person.
-        """
-        return self._org_role_join().filter(
-            or_(OrgRole.person == person,
-                ContactParticipation.contact == person))
-
-    def _org_role_join(self):
-        return self.outerjoin(
-            OrgRole, OrgRoleParticipation.org_role_id == OrgRole.org_role_id)
-
-
-class ContactParticipationQuery(ParticipationQuery):
-
-    def by_participant(self, contact):
-        return self.filter_by(contact=contact)
-
-
-class OrgRoleParticipationQuery(ParticipationQuery):
-
-    def by_participant(self, org_role):
-        return self.filter_by(org_role=org_role)
-
-
-class OgdsUserParticipationQuery(ParticipationQuery):
-
-    def by_participant(self, ogds_user):
-        return self.filter_by(ogds_userid=ogds_user.id)
-
-
 class Participation(Base, SQLFormSupport):
     """Base class for participations.
     """
-    query_cls = ParticipationQuery
     __tablename__ = 'participations'
 
     participation_id = Column('id', Integer, Sequence('participations_id_seq'),
@@ -173,7 +126,6 @@ class OgdsUserParticipation(Participation):
     include a foreign key to the ogds-user.
 
     """
-    query_cls = OgdsUserParticipationQuery
     __mapper_args__ = {'polymorphic_identity': 'ogds_user_participation'}
 
     ogds_userid = Column(String(USER_ID_LENGTH))
@@ -202,14 +154,12 @@ class OgdsUserParticipation(Participation):
         attributes['ogds_userid'] = self.ogds_userid
         return attributes
 
-
 OgdsUserToContactAdapter.participation_class = OgdsUserParticipation
 
 
 class ContactParticipation(Participation):
     """Let Contacts participate in dossiers with specified roles.
     """
-    query_cls = ContactParticipationQuery
     __mapper_args__ = {'polymorphic_identity': 'contact_participation'}
 
     contact_id = Column(Integer, ForeignKey('contacts.id'))
@@ -235,7 +185,6 @@ class ContactParticipation(Participation):
 class OrgRoleParticipation(Participation):
     """Let OrgRoles participate in dossiers with specified roles.
     """
-    query_cls = OrgRoleParticipationQuery
     __mapper_args__ = {'polymorphic_identity': 'org_role_participation'}
 
     org_role_id = Column(Integer, ForeignKey('org_roles.id'))
